Allow More Than One ID Per Forge
With only a single user ID for each git forge, the script's
usefulness was fairly limited. With this commit, it now handles more
than one user ID for a given git forge.

Instead of choosing a remote, the user is now prompted for the user ID
to use. As before, the user ID associated with the remote called
'origin' is presented as the default choice.