Troublesome Default Identities Directory Name
The default name of "~/.git/id" was fairly unfortunate, since the
presence of a ".git" directory (despite it NOT containing any "config"
file) tricks many tools into thinking the home directory was a
repository or working copy.

This commit changes the name of the default identities directory to a
less controversial "~/.gitids".